requirements.txt
python-dotenv==1.0.1
langchain==0.2.6
langchain-openai==0.1.14
streamlit==1.36.0